requirements.txt
Flask
gunicorn
sklearn
pandas
numpy
matplotlib
seaborn
ipykernel
pyYAML
-e .